+++ plugins/advanced-responsive-video-embedder/advanced-responsive-video-embedder.php
@@ -3,7 +3,7 @@
  * Plugin Name:       ARVE Advanced Responsive Video Embedder
  * Plugin URI:        https://nextgenthemes.com/plugins/arve-pro/
  * Description:       Easy responsive video embeds via URL (like WordPress) or Shortcodes. Supports almost anything you can imagine.
- * Version:           9.5.14
+ * Version:           9.6.2
  * Author:            Nicolas Jonas
  * Author URI:        https://nextgenthemes.com
  * License:           GPL-3.0
@@ -19,12 +19,13 @@
 
 namespace Nextgenthemes\ARVE;
 
-const VERSION               = '9.5.14';
-const PRO_VERSION_REQUIRED  = '5.1.11';
+const VERSION               = '9.6.2';
+const PRO_VERSION_REQUIRED  = '5.2.0';
 const NUM_TRACKS            = 3;
 const PLUGIN_FILE           = __FILE__;
 const PLUGIN_DIR            = __DIR__;
 const VIDEO_FILE_EXTENSIONS = array( 'av1mp4', 'mp4', 'm4v', 'webm', 'ogv' );
+const PREMIUM_SECTIONS      = array( 'pro', 'sticky-videos', 'random-video', 'videojs' );
 const DEFAULT_MAXWIDTH      = 900;
 const OEMBED_HTML_PRIORITY  = -5;