+++ plugins/advanced-responsive-video-embedder/php/functions-html-output.php
@@ -57,7 +57,6 @@ function build_iframe_tag( array $a ) {
 	if ( 'wistia' === $a['provider'] ) {
 		$class   .= ' wistia_embed';
 		$sandbox .= ' allow-forms';
-
 	}
 
 	if ( 'zoom' === $a['provider'] ) {
@@ -83,6 +82,7 @@ function build_iframe_tag( array $a ) {
 				'scrolling'       => 'no',
 				'src'             => $a['src'],
 				'width'           => $a['width'],
+				'title'           => $a['title'],
 			),
 		),
 		$a
@@ -104,7 +104,7 @@ function build_video_tag( array $a ) {
 				'controls'           => $a['controls'],
 				'controlslist'       => $a['controlslist'],
 				'loop'               => $a['loop'],
-				'preload'            => 'metadata',
+				'preload'            => 'none',
 				'width'              => is_feed() ? $a['width'] : false,
 				'poster'             => empty( $a['img_src'] ) ? false : $a['img_src'],
 				// ARVE only
@@ -383,36 +383,27 @@ function promote_link( $arve_link ) {
 
 function arve_embed( $html, array $a ) {
 
-	$class      = 'arve-embed';
-	$ratio_span = '';
+	$class = 'arve-embed';
+	$style = false;
 
 	if ( $a['aspect_ratio'] ) {
 		$class .= ' arve-embed--has-aspect-ratio';
 	}
 
-	if ( '16:9' === $a['aspect_ratio'] ) {
-		$class .= ' arve-embed--16by9';
-	} elseif ( $a['aspect_ratio'] ) {
-		$ratio_span = build_tag(
-			array(
-				'name'       => 'ar',
-				'tag'        => 'span', // so we output it within <p>
-				'inner_html' => '',
-				'attr'       => array(
-					'class' => 'arve-ar',
-					'style' => sprintf( 'padding-top:%F%%', aspect_ratio_to_percentage( $a['aspect_ratio'] ) ),
-				),
-			),
-			$a
-		);
+	if ( ! in_array($a['aspect_ratio'], [ '16:9', '375:211' ], true) ) {
+		$ar    = str_replace( ':', ' / ', $a['aspect_ratio'] );
+		$style = sprintf( 'aspect-ratio: %s', $ar );
 	}
 
 	return build_tag(
 		array(
 			'name'       => 'embed',
 			'tag'        => 'span', // so we output it within <p>
-			'inner_html' => $ratio_span . $html,
-			'attr'       => array( 'class' => $class ),
+			'inner_html' => $html,
+			'attr'       => array(
+				'class' => $class,
+				'style' => $style,
+			),
 		),
 		$a
 	);