+++ plugins/advanced-responsive-video-embedder/src/admin.ts
@@ -1,4 +1,5 @@
 import './admin.scss';
+import { initSC } from './admin-shortcode-creator';
 
 export {};
 declare global {
@@ -6,31 +7,9 @@ declare global {
 		wp;
 		jQuery;
 		sui;
-		/* eslint-disable-next-line */
-		tb_show;
 		ajaxurl;
 	}
 }
-
-window.jQuery(document).on('click', '#arve-btn', function () {
-	const sui = window.sui;
-
-	if ('undefined' !== typeof sui) {
-		sui.utils.shortcodeViewConstructor.parseShortcodeString('[arve]');
-
-		window.wp
-			.media({
-				frame: 'post',
-				state: 'shortcode-ui',
-				currentShortcode: sui.utils.shortcodeViewConstructor.parseShortcodeString(
-					'[arve]'
-				),
-			})
-			.open();
-	} else {
-		window.tb_show('ARVE Optional Features', '#TB_inline?inlineId=arve-thickbox');
-	}
-});
 
 window.jQuery(document).ready(function ($) {
 	$('.notice.is-dismissible').on('click', '.notice-dismiss', function (event) {
@@ -46,3 +25,5 @@ window.jQuery(document).ready(function ($) {
 		});
 	});
 });
+
+initSC();