+++ plugins/advanced-responsive-video-embedder/tests/tests-urls.php
@@ -7,7 +7,7 @@ class Tests_URLs extends WP_UnitTestCase {
 
 		$output = apply_filters( 'the_content', 'https://vimeo.com/265932452' );
 
-		$this->assertNotContains( 'Error', $output );
-		$this->assertContains( 'src="https://player.vimeo.com/video/265932452', $output );
+		$this->assertStringNotContainsString( 'Error', $output );
+		$this->assertStringContainsString( 'src="https://player.vimeo.com/video/265932452', $output );
 	}
 }